pyproject.toml
[project]
name = "oil-agent"
version = "1.0.0"
description = "Agent de surveillance du marché pétrolier basé sur smolagents + LiteLLM + Ollama"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Oil Monitor Team"}
]
dependencies = [
    "smolagents[litellm]",
    "duckduckgo-search",
    "ddgs",
    "requests",
    "beautifulsoup4",
    "markdownify",
    "feedparser",
    "dspy-ai>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "ruff",
]
